Implement UUID v7 primary keys for Owner and Expense models; update r… (#20)
* Implement UUID v7 primary keys for Owner and Expense models; update related services and migrations

- Replace auto-increment integer primary keys with UUID v7 for `users` and `expenses` tables to support offline capture and enhance identity security.
- Introduce `HasUuids` trait in `User` and `Expense` models, ensuring consistent UUID handling.
- Update `ExpenseAggregateCache` methods to accept string UUIDs instead of integers for owner IDs.
- Create an in-place migration to backfill existing data and enforce foreign key constraints on `expenses.user_id`.
- Clear sessions and flush cache during migration to prevent stale data references.
- Add tests to validate UUID implementation and migration integrity.


* Update CreateUserCommand to display resolved user name; enhance UUID migration logic

- Modify CreateUserCommand to output the resolved user name instead of the email upon successful creation.
- Update migration to ensure 'user_id' in 'sessions' table is nullable and change its type to UUID if it exists.
- Introduce a new migration to ensure domain tables utilize UUID v7 primary keys, with a safeguard against reversal without a backup.
- Adjust tests to reflect changes in user creation output and validate UUID primary key implementation across relevant tables.

---------

+++ resources/views/pages/⚡home.blade.php
@@ -94,7 +94,7 @@ new class extends Component
     private function calculateTotal(): void
     {
         $this->total = $this->aggregateCache()->dayTotal(
-            (int) auth()->id(),
+            auth()->id(),
             Carbon::parse($this->spent_on)
         );
     }
@@ -102,7 +102,7 @@ new class extends Component
     private function calculateAverage(): void
     {
         $dailyTotals = collect($this->aggregateCache()->monthDayTotals(
-            (int) auth()->id(),
+            auth()->id(),
             now()
         ));
